linux-at91: 3.18: only deal with .its if they are present

+++ recipes-kernel/linux/linux-at91_3.18.bb
@@ -29,11 +29,13 @@ do_deploy_append() {
 			DTB_PATH="${B}/arch/${ARCH}/boot/"
 		fi
 
-		cp ${S}/arch/${ARCH}/boot/dts/${MACHINE}*.its ${DTB_PATH}
-		cd ${DTB_PATH}
-		mkimage -f ${MACHINE}.its ${MACHINE}.itb
-		install -m 0644 ${MACHINE}.itb ${DEPLOYDIR}/${MACHINE}.itb
-		cd -
+		if [ -e ${S}/arch/${ARCH}/boot/dts/${MACHINE}.its ]; then
+			cp ${S}/arch/${ARCH}/boot/dts/${MACHINE}*.its ${DTB_PATH}
+			cd ${DTB_PATH}
+			mkimage -f ${MACHINE}.its ${MACHINE}.itb
+			install -m 0644 ${MACHINE}.itb ${DEPLOYDIR}/${MACHINE}.itb
+			cd -
+		fi
 	fi
 }
